fix(money,bill): discount cash ใช้ float64 และเลขบิลนับจากเลขสูงสุดแทน COUNT
สองข้อจากรายการที่ค้าง

1. discount cash เปลี่ยน float32 เป็น float64 ทั้ง 20 จุด

   ข้อ 10 ใน SECURITY_AUDIT เขียนว่ายอดที่ควรเป็น 0 จะไปจบที่ -0.0000019
   ซึ่งเกิดในฐานข้อมูลไม่ได้ ตรวจ 19 ก.ย. พบคอลัมน์ balance_discount_cash
   กับ discount_cash เป็น numeric(8,2) ปัดสองตำแหน่งทุกครั้งที่เขียน
   และข้อมูลจริง 2,704 แถวไม่มีแถวไหนติดลบหรือมีเศษต่ำกว่า 0.01 เลย

   ความเสี่ยงจริงอยู่ฝั่ง Go: float32 เก็บเลขนัยสำคัญได้ราว 7 หลัก
   แต่คอลัมน์รับได้ถึง 999999.99 ซึ่งเป็น 8 หลัก ค่าที่ใกล้เพดานจึงเพี้ยน
   ตั้งแต่ตอนแปลงชนิด ไม่ใช่ตอนลบซ้ำ ๆ อย่างที่รายงานเดิมเขียน

2. GenerateBillNo อ่านเลขสูงสุดที่เคยออกแล้วแทนการ COUNT แถว

   COUNT มีจุดอ่อนสองอย่าง ถ้าลบบิลตัวนับจะถอยไปทับเลขที่ออกไปแล้ว
   และมันนับบิลของ pos นั้นทุกใบไม่ว่ามาจาก stack ไหน เลขของแต่ละ stack
   จึงกระโดดตามปริมาณงานของอีกฝั่ง เปลี่ยนมาหาค่าสูงสุดจากเลขบิลที่ขึ้นต้น
   ด้วย prefix ของตัวเอง แก้ทั้งสองข้อและไม่ต้องพึ่ง bill_date อีก

   ทดสอบกับข้อมูลจริง 3 เคส: วันที่มี 1 ใบได้ 2, วันที่มีเลขซ้ำ 0003 ได้ 4
   (COUNT จะให้ 5 แล้วข้าม 0004), วัน/prefix ใหม่ได้ 1

   ที่ยังไม่ปิดคือสองคำขอพร้อมกันอ่านเลขเดียวกัน ต้องมี unique index บน
   bill_no ซึ่งเป็น schema ของฐานที่ใช้ร่วมกับระบบเก่า และข้อมูลตอนนี้มี
   เลขซ้ำอยู่ 23 กลุ่มจึงสร้าง index ไม่ผ่าน ต้องเป็นการตัดสินใจแยก

ตรวจแล้ว go build + go vet ผ่าน exit 0, gofmt ไม่แย่ลงจาก baseline

+++ controllers/card_entity_controller.go
@@ -476,8 +476,8 @@ func TopupCardPOS(c *gin.Context) {
 					return nil
 				}(),
 				//TODO : discount cash
-				DiscountCash:        float32(posMenu.DiscountCash),
-				BalanceDiscountCash: float32(posMenu.DiscountCash),
+				DiscountCash:        float64(posMenu.DiscountCash),
+				BalanceDiscountCash: float64(posMenu.DiscountCash),
 				DiscountCashExpire: func() *time.Time {
 					if posMenu.DiscountCashLimitDays != nil && *posMenu.DiscountCashLimitDays > 0 {
 						convertDiscountCashExpire := utils.AddDaysBangkok(*posMenu.DiscountCashLimitDays)
